requirements.txt
requests
beautifulsoup4
torch
transformers
sentence-transformers
langchain
langchain-community
langchain-huggingface
faiss-cpu
chromadb  
fastapi
uvicorn
numpy
pandas
ipykernel
ipywidgets
streamlit
python-dotenv
openai
langchain-openai
#-e .